% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpile-package.R
\docType{package}
\name{ggpile-package}
\alias{ggpile}
\alias{ggpile-package}
\title{ggpile: Make Piled Bar Charts with 'ggplot2'}
\description{
Use piled 'ggplot2' scales and position to create piled bar charts with custom geoms.
}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

}
\keyword{internal}
